Move tests to TCK
